Bypass the cache
As the local cache is indexed by URL, every new request keeps displaying the same quote. We need to make them unique. 
A cross-browser compatible approach to bypassing the cache is appending a timestamp to the URL, being sure to include a "?" or "&" as appropriate. This causes every request to be unique, thereby bypassing the cache.

https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/Using_XMLHttpRequest

We are not going to use: 

$.ajaxSetup({ cache: false });

since that will disable the caching globally and there may be other calls where we would like to cache. Therefore, this uniqueness pattern allows us to bypass the cache for this single call.